+++ config/test.exs
@@ -3,11 +3,12 @@ import Config
 alias Commanded.EventStore.Adapters.InMemory
 alias Commanded.Serialization.JsonSerializer
 
+config :logger, level: :debug
 config :logger, :console, level: :debug, format: "[$level] $message\n"
 
 config :ex_unit,
   assert_receive_timeout: 1_000,
-  capture_log: true,
+  capture_log: [level: :debug],
   exclude: [:distributed]
 
 config :commanded,