+++ lib/commanded/event/handler.ex
@@ -484,7 +484,7 @@ defmodule Commanded.Event.Handler do
           case Map.get(context, :failures) do
             too_many when too_many >= 3 ->
               # skip bad event after third failure
-              Logger.warn(fn -> "Skipping bad event, too many failures: " <> inspect(event) end)
+              Logger.warning("Skipping bad event, too many failures: " <> inspect(event))
 
               :skip
 
@@ -730,7 +730,7 @@ defmodule Commanded.Event.Handler do
 
   @impl GenServer
   def terminate(reason, state) do
-    Logger.debug(fn -> describe(state) <> " is shutting down due to #{inspect(reason)}" end)
+    Logger.debug(describe(state) <> " is shutting down due to #{inspect(reason)}")
   end
 
   @doc false
@@ -756,10 +756,10 @@ defmodule Commanded.Event.Handler do
         end
 
       {:stop, reason} ->
-        Logger.debug(fn ->
+        Logger.debug(
           describe(state) <>
             " `before_reset/0` callback has requested to stop. (reason: #{inspect(reason)})"
-        end)
+        )
 
         {:stop, reason, state}
     end
@@ -778,7 +778,7 @@ defmodule Commanded.Event.Handler do
         {:subscribed, subscription},
         %Handler{subscription: %Subscription{subscription_pid: subscription}} = state
       ) do
-    Logger.debug(fn -> describe(state) <> " has successfully subscribed to event store" end)
+    Logger.debug(describe(state) <> " has successfully subscribed to event store")
 
     %Handler{handler_module: handler_module} = state
 
@@ -787,7 +787,7 @@ defmodule Commanded.Event.Handler do
         {:noreply, state}
 
       {:stop, reason} ->
-        Logger.debug(fn -> describe(state) <> " `init/0` callback has requested to stop" end)
+        Logger.debug(describe(state) <> " `init/0` callback has requested to stop")
 
         {:stop, reason, state}
     end
@@ -798,7 +798,7 @@ defmodule Commanded.Event.Handler do
   def handle_info({:events, events}, %Handler{} = state) do
     %Handler{application: application} = state
 
-    Logger.debug(fn -> describe(state) <> " received events: #{inspect(events)}" end)
+    Logger.debug(describe(state) <> " received events: #{inspect(events)}")
 
     try do
       state =
@@ -820,7 +820,7 @@ defmodule Commanded.Event.Handler do
         {:DOWN, ref, :process, _pid, reason},
         %Handler{subscription: %Subscription{subscription_ref: ref}} = state
       ) do
-    Logger.debug(fn -> describe(state) <> " subscription DOWN due to: #{inspect(reason)}" end)
+    Logger.debug(describe(state) <> " subscription DOWN due to: #{inspect(reason)}")
 
     # Stop event handler when event store subscription process terminates.
     {:stop, reason, state}
@@ -828,10 +828,22 @@ defmodule Commanded.Event.Handler do
 
   @doc false
   @impl GenServer
+  def handle_info({:EXIT, _pid, :normal}, state) do
+    # linked process exited normally, don't shutdown
+    {:noreply, state}
+  end
+
+  @impl GenServer
+  def handle_info({:EXIT, _pid, reason}, state) do
+    {:stop, reason, state}
+  end
+
+  @doc false
+  @impl GenServer
   def handle_info(message, state) do
-    Logger.error(fn ->
+    Logger.error(
       describe(state) <> " received unexpected message: " <> inspect(message, pretty: true)
-    end)
+    )
 
     {:noreply, state}
   end
@@ -854,11 +866,11 @@ defmodule Commanded.Event.Handler do
       {:error, error} ->
         {backoff, subscription} = Subscription.backoff(subscription)
 
-        Logger.info(fn ->
+        Logger.info(
           describe(state) <>
             " failed to subscribe to event store due to: " <>
             inspect(error) <> ", retrying in " <> inspect(backoff) <> "ms"
-        end)
+        )
 
         subscribe_timer = Process.send_after(self(), :subscribe_to_events, backoff)
 
@@ -875,7 +887,7 @@ defmodule Commanded.Event.Handler do
          %Handler{last_seen_event: last_seen_event} = state
        )
        when not is_nil(last_seen_event) and event_number <= last_seen_event do
-    Logger.debug(fn -> describe(state) <> " has already seen event ##{inspect(event_number)}" end)
+    Logger.debug(describe(state) <> " has already seen event ##{inspect(event_number)}")
 
     confirm_receipt(event, state)
   end
@@ -910,7 +922,7 @@ defmodule Commanded.Event.Handler do
         handle_event_error(error, event, failure_context, state)
 
       {:error, reason, stacktrace} ->
-        log_event_error({:error, reason}, event, state)
+        log_event_error({:error, reason, stacktrace}, event, state)
         telemetry_exception(start_time, :error, reason, stacktrace, telemetry_metadata)
 
         failure_context = build_failure_context(event, context, stacktrace, state)
@@ -918,14 +930,14 @@ defmodule Commanded.Event.Handler do
         handle_event_error({:error, reason}, event, failure_context, state)
 
       invalid ->
-        Logger.error(fn ->
+        Logger.error(
           describe(state) <>
             " failed to handle event " <>
             inspect(event, pretty: true) <>
             ", `handle/2` function returned an invalid value: " <>
             inspect(invalid, pretty: true) <>
             ", expected `:ok` or `{:error, term}`"
-        end)
+        )
 
         telemetry_stop(start_time, Map.put(telemetry_metadata, :error, :invalid_return_value))
 
@@ -947,8 +959,6 @@ defmodule Commanded.Event.Handler do
     rescue
       error ->
         stacktrace = __STACKTRACE__
-        Logger.error(fn -> Exception.format(:error, error, stacktrace) end)
-
         {:error, error, stacktrace}
     end
   end
@@ -1001,22 +1011,20 @@ defmodule Commanded.Event.Handler do
     case handler_module.error(error, data, failure_context) do
       {:retry, %FailureContext{context: context}} when is_map(context) ->
         # Retry the failed event
-        Logger.info(fn -> describe(state) <> " is retrying failed event" end)
+        Logger.info(describe(state) <> " is retrying failed event")
 
         handle_event(failed_event, context, state)
 
       {:retry, context} when is_map(context) ->
         # Retry the failed event
-        Logger.info(fn -> describe(state) <> " is retrying failed event" end)
+        Logger.info(describe(state) <> " is retrying failed event")
 
         handle_event(failed_event, context, state)
 
       {:retry, delay, %FailureContext{context: context}}
       when is_map(context) and is_integer(delay) and delay >= 0 ->
         # Retry the failed event after waiting for the given delay, in milliseconds
-        Logger.info(fn ->
-          describe(state) <> " is retrying failed event after #{inspect(delay)}ms"
-        end)
+        Logger.info(describe(state) <> " is retrying failed event after #{inspect(delay)}ms")
 
         :timer.sleep(delay)
 
@@ -1024,9 +1032,7 @@ defmodule Commanded.Event.Handler do
 
       {:retry, delay, context} when is_map(context) and is_integer(delay) and delay >= 0 ->
         # Retry the failed event after waiting for the given delay, in milliseconds
-        Logger.info(fn ->
-          describe(state) <> " is retrying failed event after #{inspect(delay)}ms"
-        end)
+        Logger.info(describe(state) <> " is retrying failed event after #{inspect(delay)}ms")
 
         :timer.sleep(delay)
 
@@ -1034,34 +1040,38 @@ defmodule Commanded.Event.Handler do
 
       :skip ->
         # Skip the failed event by confirming receipt
-        Logger.info(fn -> describe(state) <> " is skipping event" end)
+        Logger.info(describe(state) <> " is skipping event")
 
         confirm_receipt(failed_event, state)
 
       {:stop, reason} ->
-        Logger.warn(fn -> describe(state) <> " has requested to stop: #{inspect(reason)}" end)
+        Logger.warning(describe(state) <> " has requested to stop: #{inspect(reason)}")
 
         # Stop event handler with given reason
         throw({:error, reason})
 
       invalid ->
-        Logger.warn(fn ->
+        Logger.warning(
           describe(state) <> " returned an invalid error response: #{inspect(invalid)}"
-        end)
+        )
 
         # Stop event handler with original error
         throw(error)
     end
   end
 
-  defp log_event_error({:error, reason}, %RecordedEvent{} = failed_event, %Handler{} = state) do
-    Logger.error(fn ->
+  defp log_event_error(error, %RecordedEvent{} = failed_event, %Handler{} = state) do
+    reason =
+      case error do
+        {:error, reason} -> inspect(reason, pretty: true)
+        {:error, reason, stacktrace} -> Exception.format(:error, reason, stacktrace)
+      end
+
+    Logger.error(
       describe(state) <>
-        " failed to handle event " <>
-        inspect(failed_event, pretty: true) <>
-        " due to: " <>
-        inspect(reason, pretty: true)
-    end)
+        " failed to handle event:\n" <>
+        inspect(failed_event, pretty: true) <> ", due to:\n" <> reason
+    )
   end
 
   # Confirm receipt of event
@@ -1075,9 +1085,7 @@ defmodule Commanded.Event.Handler do
 
     %RecordedEvent{event_number: event_number} = event
 
-    Logger.debug(fn ->
-      describe(state) <> " confirming receipt of event ##{inspect(event_number)}"
-    end)
+    Logger.debug(describe(state) <> " confirming receipt of event ##{inspect(event_number)}")
 
     :ok = Subscription.ack_event(subscription, event)
     :ok = Subscriptions.ack_event(application, handler_name, consistency, event)
@@ -1110,7 +1118,7 @@ defmodule Commanded.Event.Handler do
     rescue
       error ->
         stacktrace = __STACKTRACE__
-        Logger.error(fn -> Exception.format(:error, error, stacktrace) end)
+        Logger.error(Exception.format(:error, error, stacktrace))
 
         1
     end